pyproject.toml
[project]
name = "bing-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scrapy-user-agents>=0.1.1",
    "scrapy>=2.12.0",
    "aiohttp>=3.11.9",
]
